director commands now working at basic level

+++ xtremvaders/src/main/java/xtremvaders/Directors/GameState.java
@@ -5,13 +5,13 @@ public class GameState {
     public int missilesOnScreen = 0;
     public boolean bossActive = false;
 
-    public int timeSinceLastWave = 0; // en secondes
-    public int timeSinceLastBoss = 0; // en secondes
+    public long timeSinceLastWave = 0; // en secondes
+    public long timeSinceLastBoss = 0; // en secondes
 
     public int asteroidsOnScreen = 0;
     public int powerUpsOnScreen = 0;
 
-    public int totalElapsedTime = 0; // Temps total depuis début partie, en secondes
+    public long totalElapsedTime = 0; // Temps total depuis début partie, en secondes
 
     public void tick() {
         timeSinceLastWave++;
